ml-service/requirements.txt
Flask==3.0.3
face-recognition==1.3.0
opencv-python==4.10.0.84
numpy==2.1.2
Pillow==10.4.0
gunicorn==23.0.0
